fix(models): обновлены типы в моделях

+++ models/attachment.py
@@ -3,5 +3,5 @@ from pydantic import BaseModel
 
 class AttachmentModel(BaseModel):
     name: str
-    size: tuple[int, int, int]
+    size: tuple[float, float, float]
     image_url: str
